requirements.txt
click==8.1.3
requests==2.28.1
python-dotenv==0.21.1
apscheduler==3.9.1